third_party/blink/web_tests/fast/dom/timer-throttling-background-page-near-alignment-interval-expected.txt

Bug 259680: JavaScript setInterval lagging behind

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS numFires is within 1 of 4
PASS successfullyParsed is true

TEST COMPLETE


This test ensures that intervals on background pages whose duration is close to the timer alignment interval don't miss every other firing.
